Refactor middleware IO headers
