+++ src/middleware/io/line_based/common.h
