+++ src/middleware/io/line_based/serial.h
@@ -21,11 +21,13 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Goby.  If not, see <http://www.gnu.org/licenses/>.
 
-#ifndef SerialMAVLink20190719H
-#define SerialMAVLink20190719H
+#ifndef SerialLineBased20190718H
+#define SerialLineBased20190718H
 
-#include "mavlink_common.h"
-#include "serial_interface.h"
+#include <regex>
+
+#include "goby/middleware/io/detail/serial_interface.h"
+#include "goby/middleware/io/line_based/common.h"
 
 namespace goby
 {
@@ -33,36 +35,37 @@ namespace middleware
 {
 namespace io
 {
-template <const goby::middleware::Group& line_in_group,
-          const goby::middleware::Group& line_out_group,
-          PubSubLayer publish_layer = PubSubLayer::INTERPROCESS,
-          PubSubLayer subscribe_layer = PubSubLayer::INTERTHREAD>
-using SerialThreadMAVLinkBase =
-    IOThreadMAVLink<line_in_group, line_out_group, publish_layer, subscribe_layer,
-                    SerialThread<line_in_group, line_out_group, publish_layer, subscribe_layer>,
-                    goby::middleware::protobuf::SerialConfig>;
-
-/// \brief Reads/Writes MAVLink message packages from/to serial port
+/// \brief Reads/Writes strings from/to serial port using a line-based (typically ASCII) protocol with a defined end-of-line regex.
 /// \tparam line_in_group goby::middleware::Group to publish to after receiving data from the serial port
 /// \tparam line_out_group goby::middleware::Group to subcribe to for data to send to the serial port
 template <const goby::middleware::Group& line_in_group,
           const goby::middleware::Group& line_out_group,
           PubSubLayer publish_layer = PubSubLayer::INTERPROCESS,
           PubSubLayer subscribe_layer = PubSubLayer::INTERTHREAD>
-class SerialThreadMAVLink
-    : public SerialThreadMAVLinkBase<line_in_group, line_out_group, publish_layer, subscribe_layer>
+class SerialThreadLineBased
+    : public detail::SerialThread<line_in_group, line_out_group, publish_layer, subscribe_layer>
 {
+    using Base =
+        detail::SerialThread<line_in_group, line_out_group, publish_layer, subscribe_layer>;
+
   public:
-    SerialThreadMAVLink(const goby::middleware::protobuf::SerialConfig& config)
-        : SerialThreadMAVLinkBase<line_in_group, line_out_group, publish_layer, subscribe_layer>(
-              config)
+    /// \brief Constructs the thread.
+    /// \param config A reference to the Protocol Buffers config read by the main application at launch
+    /// \param index Thread index for multiple instances in a given application (-1 indicates a single instance)
+    SerialThreadLineBased(const goby::middleware::protobuf::SerialConfig& config, int index = -1)
+        : Base(config, index), eol_matcher_(this->cfg().end_of_line())
     {
     }
 
-    ~SerialThreadMAVLink() {}
+    ~SerialThreadLineBased() {}
 
   private:
+    /// \brief Starts an asynchronous read on the serial port until the end-of-line string is reached. When the read completes, a lambda is called that publishes the received line.
     void async_read() override;
+
+  private:
+    match_regex eol_matcher_;
+    boost::asio::streambuf buffer_;
 };
 } // namespace io
 } // namespace middleware
@@ -72,16 +75,18 @@ template <const goby::middleware::Group& line_in_group,
           const goby::middleware::Group& line_out_group,
           goby::middleware::io::PubSubLayer publish_layer,
           goby::middleware::io::PubSubLayer subscribe_layer>
-void goby::middleware::io::SerialThreadMAVLink<line_in_group, line_out_group, publish_layer,
-                                               subscribe_layer>::async_read()
+void goby::middleware::io::SerialThreadLineBased<line_in_group, line_out_group, publish_layer,
+                                                 subscribe_layer>::async_read()
 {
-    boost::asio::async_read(
-        this->mutable_serial_port(), boost::asio::buffer(this->buffer()),
-        boost::asio::transfer_at_least(1),
+    boost::asio::async_read_until(
+        this->mutable_serial_port(), buffer_, eol_matcher_,
         [this](const boost::system::error_code& ec, std::size_t bytes_transferred) {
             if (!ec && bytes_transferred > 0)
             {
-                this->try_parse(bytes_transferred);
+                std::string bytes(bytes_transferred, 0);
+                std::istream is(&buffer_);
+                is.read(&bytes[0], bytes_transferred);
+                this->handle_read_success(bytes_transferred, bytes);
                 this->async_read();
             }
             else
